+++ Sources/Authentication/Authenticatable.swift
@@ -6,54 +6,7 @@ import Vapor
 /// like PasswordAuthenticatable.
 public protocol Authenticatable: Model { }
 
-final class AuthenticationCache {
-    private var storage: [ObjectIdentifier: Any]
 
-    init() {
-        self.storage = [:]
-    }
-
-    internal subscript<A>(_ type: A.Type) -> A?
-        where A: Authenticatable
-    {
-        get { return storage[ObjectIdentifier(A.self)] as? A }
-        set { storage[ObjectIdentifier(A.self)] = newValue }
-    }
-}
-
-extension Request {
-    public func authenticate<A>(_ instance: A) throws
-        where A: Authenticatable
-    {
-        let cache = try make(AuthenticationCache.self)
-        cache[A.self] = instance
-    }
-
-    public func authenticated<A>(_ type: A.Type) throws -> A?
-        where A: Authenticatable
-    {
-        let cache = try make(AuthenticationCache.self)
-        return cache[A.self]
-    }
-
-    public func isAuthenticated<A>(_ type: A.Type) throws -> Bool
-        where A: Authenticatable
-    {
-        return try authenticated(A.self) != nil
-    }
-
-    public func requireAuthenticated<A>(_ type: A.Type) throws -> A
-        where A: Authenticatable
-    {
-        guard let a = try authenticated(A.self) else {
-            throw AuthenticationError(
-                identifier: "notAuthenticated",
-                reason: "\(A.self) has not been authenticated."
-            )
-        }
-        return a
-    }
-}
 
 import Crypto
