+++ Tests/ServiceTests/Utilities.swift
@@ -17,7 +17,7 @@ class PrintLog: Log {
 extension PrintLog: ServiceType {
     static let serviceName = "print"
     static let serviceSupports: [Any.Type] = [Log.self]
-    static func makeService(for container: Container) throws -> Self? {
+    static func makeService(for container: Container) throws -> Self {
         return .init()
     }
 }
@@ -32,7 +32,7 @@ class AllCapsLog: Log {
 extension AllCapsLog: ServiceType {
     static let serviceName = "all-caps"
     static let serviceSupports: [Any.Type] = [Log.self]
-    static func makeService(for container: Container) throws -> Self? {
+    static func makeService(for container: Container) throws -> Self {
         return .init()
     }
 }
@@ -76,7 +76,7 @@ extension BCryptHasher: ServiceType {
         return [Hasher.self]
     }
 
-    static func makeService(for container: Container) throws -> Self? {
+    static func makeService(for container: Container) throws -> Self {
         let config = try container.make(BCryptConfig.self, for: BCryptHasher.self)
         return .init(cost: config.cost)
     }
@@ -100,7 +100,7 @@ extension BCryptConfig: ServiceType {
         return []
     }
 
-    static func makeService(for container: Container) throws -> BCryptConfig? {
+    static func makeService(for container: Container) throws -> BCryptConfig {
         let cost: Int
 
         switch container.environment {
